migrating to cFramework

+++ middlewareVision/src/middlewareVision/nodes/External/Alert.java
@@ -1,6 +1,7 @@
 package middlewareVision.nodes.External;
 
-import kmiddle2.nodes.areas.Area;
+
+import cFramework.nodes.area.Area;
 import middlewareVision.config.AreaNames;
 
 /**
@@ -13,7 +14,6 @@ public class Alert extends Area {
         this.ID = AreaNames.Alert;
         this.namer = AreaNames.class;
         addProcess(AlertProcess.class);
-
         //@AddProcess
     }
 
@@ -23,7 +23,7 @@ public class Alert extends Area {
     }
 
     @Override
-    public void receive(int nodeID, byte[] data) {
+    public void receive(long nodeID, byte[] data) {
         send(AreaNames.AlertProcess, data);
         //@SendProcess
     }
