+++ middlewareVision/src/middlewareVision/nodes/External/TestAttention.java
@@ -2,7 +2,8 @@ package middlewareVision.nodes.External;
 
 
 
-import kmiddle2.nodes.areas.Area;
+
+import cFramework.nodes.area.Area;
 import middlewareVision.config.AreaNames;
 
 
@@ -24,7 +25,7 @@ public class TestAttention extends Area{
     }
 
     @Override
-    public void receive(int nodeID, byte[] data) {
+    public void receive(long nodeID, byte[] data) {
         send(AreaNames.FeedbackProccess,data);
 	
 //@SendProcess
