+++ middlewareVision/src/middlewareVision/nodes/Visual/MST/MST.java
@@ -1,38 +1,36 @@
 package middlewareVision.nodes.Visual.MST;
 
-
-
-import kmiddle2.nodes.areas.Area;
+import cFramework.nodes.area.Area;
 import middlewareVision.config.AreaNames;
-
 
 /**
  *
- * 
+ *
  */
-public class MST extends Area{
+public class MST extends Area {
+
     public MST() {
         this.ID = AreaNames.MST;
         this.namer = AreaNames.class;
         addProcess(MSTPolarCells.class);
-	
-addProcess(MSTTemplateCells.class);
-	
-	//@AddProcess
+
+        addProcess(MSTTemplateCells.class);
+
+        //@AddProcess
     }
 
     @Override
     public void init() {
-        
+
     }
 
     @Override
-    public void receive(int nodeID, byte[] data) {
-        send(AreaNames.MSTPolarCells,data);
-	
-send(AreaNames.MSTTemplateCells,data);
-	
-	//@SendProcess
+    public void receive(long nodeID, byte[] data) {
+        send(AreaNames.MSTPolarCells, data);
+
+        send(AreaNames.MSTTemplateCells, data);
+
+        //@SendProcess
     }
-    
+
 }
