+++ middlewareVision/src/middlewareVision/nodes/Visual/MT/MT.java
@@ -1,6 +1,6 @@
 package middlewareVision.nodes.Visual.MT;
 
-import kmiddle2.nodes.areas.Area;
+import cFramework.nodes.area.Area;
 import middlewareVision.config.AreaNames;
 
 /**
@@ -23,7 +23,7 @@ public class MT extends Area {
     }
 
     @Override
-    public void receive(int nodeID, byte[] data) {
+    public void receive(long nodeID, byte[] data) {
         send(AreaNames.MTComponentCells, data);
         send(AreaNames.MTPatternCells, data);
         //@SendProcess
